backend/requirements.txt
Flask==3.1.1
pyodbc==5.2.0
python-dotenv==1.1.1